Create Ice cream flavour
 Choose your proficient language to answer.
There is a system which allows you to buy ice cream with different toppings. Based on the
toppings, the flavour and the price will be different and customers can define the order and the
number of the toppings which can be included. Design a system with appropriate design
patterns and OOP concepts to achieve this ice cream ordering system. Following are the
constraints.
1. Customer can order a plain ice cream or with any number of toppings
2. Vendor can introduce more toppings with different price additions into the system without
changing the existing system
3. Following are example toppings
a. Peanut
b. Whipped cream
c. Hard chocolate
4. When the customer makes the necessary order, the system should provide two options
to
a. Print what are the final formation of the ice cream
b. Print the final price of the ice cream